Add aug_info API function
aug_info API function provides information about the node of the specified tree
path. It sets the absolute path of the file name and indexes of label and value
inside the node. It returns 0 in case of success and -1 if the node is not
associated to a file or if the path is invalid.  The main use case is to make
possible to display the related file with the node elements hightlighted in a
UI.

  * src/augeas.c : add aug_info implementation
  * src/augeas.h : add declaration of aug_info
  * src/augeas_sym.version : make the new function visible in libaugeas
  * src/augtool.c : add the "info" command to augtool
  * src/get.c : gather filename, label and value position if any
  * src/info.c : create and release functions for struct node_info
  * src/info.h : add definition of struct node_info
  * src/internal.h : add node_info to struct tree
  * src/transform.c : duplicate filename string
  * tests/test-load.c : 3 basic unit tests for general correct behavior