Basic support for multipart keys
Multipart key is useful to avoid put ambiguity between two similar keys. In the
get direction, it adds a prefix to the key, and removes it in the put
direction.

  * src/get.c: fix get_key to concatenate previously existing label
  * src/lens.c: accept one label and one key in a concat, the label must be
    before the key, and concatenate the label to the key to form the tree
    node label
  * src/lens.h: distinguish between key and label lens
  * src/put.c: remove prefix from the tree node label
  * tests/modules/fail_multi_label_concat.aug: avoid multiple labels in subtree
  * tests/modules/fail_multipart_ambig.aug: avoid multiple keys in subtree
  * tests/modules/fail_multipart_order.aug: avoir multiple labels in subtree
  * tests/modules/pass_multipart_key.aug: unittests for correct behavior

The test pass_multipart_key.aug is known to fail in this version on a recursive
lens. It triggers a segfault from bug_on that should be diagnose. Recursive
lens is not yet properly handled.