Add support for multipart keys
Multipart keys are keys composed of multiple key and label lenses. In the get
direction, the node label is the concatenation of all keys and labels of the
subtree. In the put direction, the node label is split between each keys and
labels inside the tree. The labels are added in the get direction and removed
in the put direction.

Multipart keys require typechecking on ktype and vtype, and atype must be done
only on complete atype, compiled by subtree. This patch replaces checks for
multiple keys inside a subtree from fail to pass.

Primitive lenses have a default ktype and vtype set to the empty pattern to
make key matching correspond to the lens structure in the put direction, and
avoid corner cases in typechecking.

  * src/get.c : build the node label by concatenating keys and labels
  * src/lens.c : perform typecheck on complete atype, add typecheck on ktype
    and vtype. Add default value to ktype and vtype for all primary lens, to
    handle all typechecking situations.
  * src/lens.h : add flag to keep track of the atype is ready for typechecking
  * src/put.c : parse the node label against the ktype of a subtree child, add
    a regs to state to store the matching result and keep track of the current
    matching group
  * src/regexp.c : handle passing NULL values to various functions

This patch includes the lens debug output patch and tests in tests/modules.