Basic support for case insensitive square lens v2
Square lens must match open and close tag. In some situations, for instance
httpd, open and close tags case is allowed to be different. If the square
regexp is case insensitive, we allow case mismatch between open and close tag.
The default case for the close tag is the one from the key in the create
function, and in the put function, the original case is preserved.

As test case shows, some serious errors occurs when the lens regexp is set to
nocase. Here is the error message while running pass_square.aug:

lt-augparse: put.c:236: split_concat: Assertion `regs.start[reg] != -1' failed.

New in v2:
  * Use STRCASEEQ macro
  * No white space cleanup
  * Never set regexp->info to NULL to avoid undefined behavior