Generic square lens v2
Lens square now takes lenses instead of regular expressions as arguments, as follow:

  square left body right

The left and right lenses must accept the same language. The left lens can be
either key or del, while the right lens must be del.

This patch changes the semantic of the square lens and is thus not backward
compatible. Modules must be updated accordingly. The definition of the square
lens should now be definitive, and later addition may add support for
additionnal lenses for left and right sides.

  * builtin.c: change definition of the square lens.
  * get.c: adds a simple AST to keep track of the parse to match left and right
    side of the square. Regular lenses are handled with regexp registers.
  * put.c: the variable state->override is used to indicate to a del part of a
    square lens that it should use the provided value instead of the saved
    string in the skel or the default value.
  * lens.c: change the lns_make_square to match the new structure.
  * info.c: dump details of error if available.
  * Update httpd.aug and xml.aug witht the new square lens definition.
  * Update unit tests according to the new definition.
  * Add new debugging option cf.get.ast to dump the AST on stdout

Changes in v2:
  * protect arg with parenthesis in child_first and child_last macros
  * drop flag in_square, avoid lens copy, drop square_left/right strings
    replaced by ast for rec lens and regs for regular lens. This
    processing is more general than the previous approach, and thus
    will help to extends the square lens more in the future.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>